lib: Make set_fds_cloexec() more strictly async-signal-safe
ioctl() technically isn't async-signal-safe. It would probably be *OK*
to call it in this case, but the docs also say that the fcntl() method
is "preferred".

So let's just always use fcntl(). Most of the time we'll probably only
need one syscall anyway, so it's not a huge perfomance loss.