fditer/dirfd: Improve accuracy of size_hint()
Now it iterates through the buffered entries, just like next(), and
tries to come up with a lower boundary. (It's not possible to determine
an upper boundary; at least not if there might be any file descriptors
left.)

The lower limit will always be returned as 0 if next() hasn't been
called yet, and it will frequently be an underestimate, but it may
improve efficiency when collect()ing a list of file descriptors.

This entailed some minor refactoring of next() to avoid duplicating some
of the pointer magic used to extract a RawDirent.