close: Add translation step to simplify minfd/keep_fds specification
If keep_fds is sorted, sometimes we can remove some of the first few
elements (possibly incrementing minfd along the way) such that the
resulting keep_fds/minfd combination has the same result but may be more
efficient.

For example, close_open_fds(3, &[3, 4]) is equivalent to
close_open_fds(5, &[]). On the BSDs, performing this simplification
allows us to jump straight to a call to closefrom(), without ever
calling iter_fds(). On other systems, it may save a few fcntl() calls.